Add some JUnit tests

+++ src/main/java/duke/command/DeleteCommand.java
@@ -17,7 +17,11 @@ public class DeleteCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        if (taskIdx > tasks.size()) {
+            throw new DukeException("No task with this ID!");
+        }
+        
         Task task = tasks.getTask(taskIdx);
         tasks.deleteTask(task);
         storage.deleteTask(task);
